reliability/requirements.txt
pyyaml==5.4.1
requests==2.25.1
aiohttp==3.7.4.post0